+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+### **0.1.0**
+
+- Retitled command quoteListInput from 'Quote List: Input Quote' to 'Quote List by Input'.
+- Added input and output separator inputBoxes for quoteListInput.
+- Added command quoteList, titled 'Quote List'. 
+- Added keybindings to call quoteList with args to replicate all removed commands/keybindings.
+- Added configuration properties: quote, inputSeparator, outputSeparator, fallbackSeparators.
+- quoteList uses fallbackSeparators if selection does not contain inputSeparator.
+- Removed commands and associated keybindings: quoteListPrimary, quoteListSecondary, quoteListTertiary, quoteListCustom.
+- Removed configuration properties: primaryQuote, secondaryQuote, tertiaryQuote.
+- Updated readme to reflect above changes.
+
 #### **0.0.4**
 
 - Added icon and category.